ubuntu-classic-2310: change pc-kernel track to 23.10
24-hwe & 23.10 have fundamentally different lifespan, GA kernel
version, and userspace initrd ABI. Especially during the overall when
both 23.10 and 24.04 are both stable and supported.

Use 23.10 track name as this is a strictly confined kernel snap usable
for core & classic systems which are based on mantic userspace. Core
systems is hypothetical, as there is no such thing as core23.10.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>