feat: azure-core-24-amd64 models
Ubuntu Core models for Microsoft Azure. Note that the `waagent`
snap is deliberately limited to the 24/beta and 24/edge channels
(no 24/stable) as the snapcraft.yaml is under development; only the
minimal set of interfaces to communicate with the Azure control plane
have been declared [0].

Until the `azure-kernel` snap is ready, the `pc-kernel` snap will
be used instead.

Refs: [0] https://forum.snapcraft.io/t/request-for-waagent/47523